Move states: date, objectFit, setObjectFit
