+++ src/hooks/usePlaylist.ts
@@ -1,22 +1,5 @@
 import { useEffect, useState } from 'react'
-import { fileNameToMoment } from '../helpers/time'
 import { Image } from '../types'
-
-const MONTHS_BY_INDEX = [
-  null,
-  'January',
-  'February',
-  'March',
-  'April',
-  'May',
-  'June',
-  'July',
-  'August',
-  'September',
-  'October',
-  'November',
-  'December',
-]
 
 type NavigateArgsT = {
   increment: number
@@ -35,7 +18,6 @@ function usePlaylist(images: Image[]) {
   // playlist = [2, 0, 1]
   const [playlist, setPlaylist] = useState<number[] | []>([])
 
-  const [objectFit, setObjectFit] = useState<'cover' | 'contain'>('cover')
   const [dateSorting, setDateSorting] = useState<'asc' | 'desc' | 'random'>(
     'asc'
   )
@@ -57,28 +39,6 @@ function usePlaylist(images: Image[]) {
         blob: images[imageIndex].blob,
       }
     })
-  }
-
-  const getDate = () => {
-    const mainImage = getMainImage()
-
-    if (!mainImage) {
-      return ''
-    }
-
-    const fileName = mainImage.name
-
-    const momentDate = fileNameToMoment(fileName)
-    const date = momentDate.date()
-    const monthName = MONTHS_BY_INDEX[momentDate.month()]
-    const year = momentDate.year()
-
-    return `${date} ${monthName} ${year}`
-  }
-
-  const handleToggleObjectFit = () => {
-    const nextState = objectFit === 'contain' ? 'cover' : 'contain'
-    setObjectFit(nextState)
   }
 
   const handleSortDate = (override?: 'asc' | 'desc') => {
@@ -213,14 +173,11 @@ function usePlaylist(images: Image[]) {
     playlist,
     playlistCursor,
     thumbnails: getThumbnails(),
-    date: getDate(),
-    objectFit,
     dateSorting,
     keyDownHandler,
     sort: handleSortDate,
     navigate: navigatePlaylist,
     navigateToDate,
-    setObjectFit: handleToggleObjectFit,
     randomizeSort: handleShuffleClick,
   }
 }
